Set pes_per_node to maximum PPN
Cray PMI expects the pes_per_node value to be the maximum number of PEs
on any node for that command. The calculated value for this could be too
small when nranks is not a multiple of nnodes, or if different
distributions are used. Fix this by setting it to the actual maximum PPN.

Tested by checking the pes_per_node value with the PMI example program.